backend/requirements.txt
asgiref==3.3.2
Django==3.2.13
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 
requests==2.26.0
pytest==6.2.4
pytest-django==3.7.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
six==1.16.0
djoser==2.1.0
Pillow==8.4.0
drf-extra-fields==3.2.1



